+++ tests/testthat/test_06-output.R
@@ -30,6 +30,7 @@ test_that("output.R -> export -> xlsx_IIASA works", {
 })
 
 test_that("cleanup output.R", {
+  skipIfPreviousFailed()
   exportfiles <- Sys.glob(file.path("..", "..", "output", "export", "*TESTTHAT*"))
   unlink(exportfiles)
 })